+++ azext_iot/tests/iotext_test_tools.py
@@ -1,0 +1,40 @@
+# coding=utf-8
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+from contextlib import contextmanager
+import sys
+import io
+
+
+@contextmanager
+def capture_output():
+    class stream_buffer_tee(object):
+        def __init__(self):
+            self.stdout = sys.stdout
+            self.buffer = io.StringIO()
+
+        def write(self, message):
+            self.stdout.write(message)
+            self.buffer.write(message)
+
+        def flush(self):
+            self.stdout.flush()
+            self.buffer.flush()
+
+        def get_output(self):
+            return self.buffer.getvalue()
+
+        def close(self):
+            self.buffer.close()
+
+    _stdout = sys.stdout
+    buffer_tee = stream_buffer_tee()
+    sys.stdout = buffer_tee
+    try:
+        yield buffer_tee
+    finally:
+        sys.stdout = _stdout
+        buffer_tee.close()
